Enhance status reporting
ClusterStatus and ExecutorStatus are now reported as type
Hashmap<String, StatusVal>. This allows composition of different status
from different subsystems in the most basic manner. Actual stats are
maintained directly using member or local variables, so that updating is
fast. The output Hashmap is only generated when status is requested.