Internode messages now encoded via Protobuf
All internode messages are now explicitly serialized via protocol
buffers. This allows the flexibility required to support rolling
upgraded/downgrade. Previously, using auto-serialized values in msgpack,
version incompatibility would result in failed deserialization if an old
message was sent to a new node. With protobufs this is no longer an
issue.

Additionally, some other cleanup of the code was made along with the
definition of the .proto schema:

`Msg`s are now parameterized by the `UserMsg` type which allows the user
to serialize their messages however they want. `RustcEncodable` and
`RustcDecodable` are no longer required as internode messages are no
longer serialized automatically. However, for now, Pids and NodeIds
still derive these traits in order to facilitate still using the
`MsgpackSerializer`.

The `Msg` enum now separates Requests and Replies into their own
sub-enums. This should make matching by callers more thorough, and
hopefully involve less default matches.

The `ThreadHandler` Service Handler implementation was removed as it's
not very useful in practice.

CorrelationIds are no longer optional in Envelopes. This made for
brittle user code since the caller almost always wanted to use a
correlation id, but would still have to check a received message to see
if it was valid and handle errors if it didn't exist where it should.
As call to `CorrelationId::pid(from_pid)` can be used to create one
for messages where it doesn't really matter.

This code once again relies on nightly builds, as it's now using the
`TryFrom` trait which is not yet stabilized. Travis builds for stable
have been removed.

The `basic` test has been removed as it didn't really add anything not
covered by other tests.

`ClusterStatus` and `ExecutorStatus` were removed in favor of Metrics.

Envelopes no longer have a new constructor. It was too easy to mix up
the to/from fields that way. They now must be created with a literal.