Use Serde instead of RustcSerialize
Derive `Serialize` and `Deserialize` for messages instead of
`RustcEncodable` and `RustcDecodable`.

Remove unnecessary trait bounds from types. These tend to transitively
'infect' any usage and require documenting the same bounds even if not
used in a given function, type or implementation. The reason I decided
to do this in this commit was because the change to Deserialize would
require a lifetime bound, which isn't even currently used in structs
deriving Deserialize. This led to a lot of unnecessary boilerplate
showing up all over the code base. Removing the trait bounds altogether
and only using them in functions and impl blocks cleaned this up.
See this issue for more information:
https://github.com/serde-rs/serde/issues/890

For further details:
https://github.com/Manishearth/rust-clippy/issues/1689#issuecomment-296396580

Simplify the Process definition by just using a single parameter instead
of an associated type.