Ensure that there are no races in Processes
Maintain a unique id for each map entry so that if a process is removed, when it
is later descheduled, it won't accidentally be put back in the map if a new
process was created with the same pid. This also ensures that we can safely add
new processes and services to the map without overwriting existing ones, as
there is no `Entry` API as in `std::collections::HashMap`.