Begin simplification: Remove Metrics and Exec timer
This is the beginning of a simplification to rabble. The goal is to streamline
the library so only thing strictly necessary to build actor systems are being
used.

In this cleanup, unused messages around timers have been removed. Timers may be
reintroduced in a more streamlined manner later. Additionally, since it was
unused the Executor Tick from the cluster server to the executor was removed.
The process timeout test was also removed, since it's no longer feasible.

Finally, in order to allow more flexible and efficient use of metrics, it was
decided to allow each actor to implement their own metrics messages in their own
manner rather than the expensive, and somewhat useless mechanism of looping over
all metrics fields, and returning a standard format consisting of a Vec of
(String, Metric) pairs.