Simplify Envelopes and Services
Remove CorrelationId's entirely. The only real use, besides some possible
request tracking across nodes, is for ensuring client requests get routed back
to the proper client service. This was mainly for tcp based services, in which
requests would come in over a tcp server service and be forwarded to the
appropriate process or service. Requests would then be routed back using
CorrelationIds. However, this added overhead to every single message, as now a
Correlation Id had to be included in each envelope, even for messages that don't
expect responses.

Additionally, the structure of the Connection Handler was not
very ergonomic. With the release of tokio and futures, it is now much easier to
implement asynchronous servers. Therefore, rabble has now removed support for
writing server endpoints and will focus only on routing messages to actors.
Replies from server endpoints will be handled per rabble message and will take either a
oneshot or mpsc sender as part of the message so replies can be sent directly to
the executing future/task.