Make using processes more ergonomic
Add a new Terminal trait and implement it for the executor. A Terminal is the
interface between processes and the rest of the system. In a process's handle
method, instead of taking a mutable reference to a vec on which to push outgoing
envelopes, a mutable reference to a terminal is passed.

Instead of having to create envelopes containing process specific messages
wrapperd in a Rabble::Msg::User() variant, processes can now send the user
message directly via a terminal via terminal.send(to, msg). The terminal will
create an envelope internally, using the pid of the sender, and ensure it gets
routed appropriately. Note that the handle method will still receive a
Rable::Msg, because it may get rabble specific messages like Msg::Timeout(id).
However, processes no longer will be able to directly create and send rabble
specific messages. Their only interaction will be through a terminal.

In addition to reducing the amount of wrapping done by processes, the terminal
has other advanatages. It abstracts out the concrete implementation of returning
messages, so that different strategies can be attempted. This allows for
performance tuning. Specifically, the implementation currently used as the
default for processes, the ExecutorTerminal will directly send envelopes over
it's fifo channel to be routed, rather than processes pushing them onto a vec
and then requiring them to be popped off, an unnecessary overhead.

The rationale for that envelope strategy was for testing, so that we could see,
and manipulate every outgoing message without having to intercept them via
reception on a channel. But with a TestTerminal implementation, a library can
implement Terminal::Send by using a Vec if it chooses, and something else if a
caller decides. We aren't forcing test execution structure on users who want to
write a test executor and control scheduling themselves.

Terminals also allow starting and cancelling timeouts. This is much more
ergonomic than destining a 'Timeout' message for some arbitrary executor pid and
having to wrap it all in an envelope and push it onto an output vec. The process
can simply start and stop timers via terminals directly. There is no loss
determinism for testing either, because a user can decide how they want to
handle these timeouts by interecepting them in the TestTerminal.

In addition to adding terminals, Ferris is now used in the cluster server
instead of the in-crate timer_wheel which has been removed. Ferris is used for
both ping timeouts in the cluster server and process timeouts.