Make rabble single threaded
There were previously 3 threads:
 * one for the poller
 * one for cluster server
 * one for executor

However, since the decision not to make the executor concurrent, there
really is no reason to keep it separate from the cluster server, and
even less reason not to merge in the poll thread.

This should actually decrease latency and use less resources as channels
are no longer needed to communicate between cluster server and executor,
resulting in less cloning and allocations. Furthermore, by putting
everhything in a single thread, it becomes easier to bring in a
string-interner and make using pids much more efficient. We can also
change the deserialization to use zero-copy when we make the change to
use a string interner for pids.